@@ -13,12 +13,12 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^13.3.11",
-    "@angular/core": "^13.3.11",
-    "@angular/forms": "^13.3.11",
-    "@angular/platform-browser": "^13.3.11",
-    "@angular/platform-browser-dynamic": "^13.3.11",
-    "@angular/router": "^13.3.11",
+    "@angular/common": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
     "@capacitor/android": "^4.0.0",
     "@capacitor/app": "^4.0.0",
     "@capacitor/core": "^4.0.0",
@@ -29,27 +29,28 @@
     "@ckeditor/ckeditor5-build-classic": "^34.0.0",
     "@ionic/angular": "^6.2.6",
     "@ionic/storage-angular": "^3.0.6",
+    "@ngx-markdown/core": "^0.2.2",
     "@types/google.maps": "^3.51.0",
     "chart.js": "^3.5.1",
     "leaflet": "^1.9.3",
     "leaflet-geosearch": "^3.7.0",
-    "ngx-markdown": "^13.1.0",
+    "ngx-markdown": "^14.0.1",
     "rxjs": "~6.6.0",
     "swiper": "^7.4.1",
     "tslib": "^2.3.1",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.9",
-    "@angular-eslint/builder": "^13.5.0",
-    "@angular-eslint/eslint-plugin": "^13.5.0",
-    "@angular-eslint/eslint-plugin-template": "^13.5.0",
+    "@angular-devkit/build-angular": "^14.2.11",
+    "@angular-eslint/builder": "^14.4.0",
+    "@angular-eslint/eslint-plugin": "^14.4.0",
+    "@angular-eslint/eslint-plugin-template": "^14.4.0",
     "@angular-eslint/schematics": "^14.4.0",
-    "@angular-eslint/template-parser": "^13.5.0",
-    "@angular/cli": "^13.3.9",
-    "@angular/compiler": "^13.3.11",
-    "@angular/compiler-cli": "^13.3.11",
-    "@angular/language-service": "~13.3.11",
+    "@angular-eslint/template-parser": "^14.4.0",
+    "@angular/cli": "^14.2.11",
+    "@angular/compiler": "^14.3.0",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/language-service": "^14.3.0",
     "@capacitor/cli": "^4.0.0",
     "@ionic/angular-toolkit": "^7.0.0",
     "@types/jasmine": "~3.6.0",
